pyproject.toml
[tool.poetry]
name = "electricity-price"
version = "1.1.0"
description = "Simple python script to fetch energy prices from entsoe.eu and publish them to MQTT broker."
authors = ["Matti Laakso <matti.laakso@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
paho-mqtt = "^1.5.1"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
